// src/api/generated/model/choiceOption.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { ChoiceOptionMediaType } from "./choiceOptionMediaType";

export interface ChoiceOption {
  id?: string;
  text?: string;
  isCorrect?: boolean;
  feedback?: string;
  mediaUrl?: string;
  mediaType?: ChoiceOptionMediaType;
}